worker/requirements.txt
psycopg2-binary
pandas
python-dotenv
numpy<2.0.0